requirements.txt
chumpy @ git+https://github.com/mattloper/chumpy
numpy==1.22.3
yacs
joblib
scikit-image
opencv-python
imageio[ffmpeg]
matplotlib
tensorboard
smplx
progress
einops
mmcv==1.3.9
timm==0.4.9
munkres
xtcocotools>=1.8
loguru
setuptools==59.5.0
tqdm
ultralytics
pytorch3d==0.7.2